--- a/aten/src/TH/generic/THBlas.cpp
+++ b/aten/src/TH/generic/THBlas.cpp
@@ -292,8 +292,20 @@ void THBlas_(ger)(int64_t m, int64_t n, scalar_t alpha, scalar_t *x, int64_t inc
   }
 }
 
-void THBlas_(gemm)(char transa, char transb, int64_t m, int64_t n, int64_t k, scalar_t alpha, scalar_t *a, int64_t lda, scalar_t *b, int64_t ldb, scalar_t beta, scalar_t *c, int64_t ldc)
-{
+void THBlas_(gemm)(
+    char transa,
+    char transb,
+    int64_t m,
+    int64_t n,
+    int64_t k,
+    scalar_t alpha,
+    scalar_t* a,
+    int64_t lda,
+    scalar_t* b,
+    int64_t ldb,
+    scalar_t beta,
+    scalar_t* c,
+    int64_t ldc) {
   int transa_ = ((transa == 't') || (transa == 'T'));
   int transb_ = ((transb == 't') || (transb == 'T'));
 
@@ -348,37 +360,46 @@ void THBlas_(gemm)(char transa, char transb, int64_t m, int64_t n, int64_t k, sc
   }
 #endif
   {
-    int64_t i, j, l;
     if(!transa_ && !transb_)
     {
-      scalar_t *a_ = a;
-      for(i = 0; i < m; i++)
-      {
-        scalar_t *b_ = b;
-        for(j = 0; j < n; j++)
-        {
-          scalar_t sum = 0;
-          for(l = 0; l < k; l++)
-            sum += a_[l*lda]*b_[l];
-          b_ += ldb;
-	  if (beta == 0)
-	    c[j*ldc+i] = alpha*sum;
-	  else
-	    c[j*ldc+i] = beta*c[j*ldc+i]+alpha*sum;
+      if (beta == 0) {
+        for (int64_t j = 0; j < n; j++) {
+          memset(c + j * ldc, 0, m);
         }
-        a_++;
+      } else {
+        for (int64_t j = 0; j < n; j++) {
+          for (int64_t i = 0; i < m; i++) {
+            c[j * ldc + i] *= beta;
+          }
+        }
+      }
+      for (int64_t l = 0; l < k; l++) {
+        for (int64_t j = 0; j < n; j++) {
+          scalar_t val = b[l + j * ldb] * alpha;
+          int64_t i_m = m / 4;
+          for (int64_t i_i = 0; i_i < i_m; i_i++) {
+            c[j * ldc + i_i * 4 + 0] += a[i_i * 4 + 0 + l * lda] * val;
+            c[j * ldc + i_i * 4 + 1] += a[i_i * 4 + 1 + l * lda] * val;
+            c[j * ldc + i_i * 4 + 2] += a[i_i * 4 + 2 + l * lda] * val;
+            c[j * ldc + i_i * 4 + 3] += a[i_i * 4 + 3 + l * lda] * val;
+          }
+          int64_t i = i_m * 4;
+          for (; i < m; i++) {
+            c[j * ldc + i] += a[i + l * lda] * val;
+          }
+        }
       }
     }
     else if(transa_ && !transb_)
     {
       scalar_t *a_ = a;
-      for(i = 0; i < m; i++)
+      for(int64_t i = 0; i < m; i++)
       {
         scalar_t *b_ = b;
-        for(j = 0; j < n; j++)
+        for(int64_t j = 0; j < n; j++)
         {
           scalar_t sum = 0;
-          for(l = 0; l < k; l++)
+          for(int64_t l = 0; l < k; l++)
             sum += a_[l]*b_[l];
           b_ += ldb;
 	  if (beta == 0)
@@ -392,13 +413,13 @@ void THBlas_(gemm)(char transa, char transb, int64_t m, int64_t n, int64_t k, sc
     else if(!transa_ && transb_)
     {
       scalar_t *a_ = a;
-      for(i = 0; i < m; i++)
+      for(int64_t i = 0; i < m; i++)
       {
         scalar_t *b_ = b;
-        for(j = 0; j < n; j++)
+        for(int64_t j = 0; j < n; j++)
         {
           scalar_t sum = 0;
-          for(l = 0; l < k; l++)
+          for(int64_t l = 0; l < k; l++)
             sum += a_[l*lda]*b_[l*ldb];
           b_++;
 	  if (beta == 0)
@@ -412,13 +433,13 @@ void THBlas_(gemm)(char transa, char transb, int64_t m, int64_t n, int64_t k, sc
     else
     {
       scalar_t *a_ = a;
-      for(i = 0; i < m; i++)
+      for(int64_t i = 0; i < m; i++)
       {
         scalar_t *b_ = b;
-        for(j = 0; j < n; j++)
+        for(int64_t j = 0; j < n; j++)
         {
           scalar_t sum = 0;
-          for(l = 0; l < k; l++)
+          for(int64_t l = 0; l < k; l++)
             sum += a_[l]*b_[l*ldb];
           b_++;
 	  if (beta == 0)