Looks like there might be some easy changes within THBlas we can make to significantly speed it up for int types. This also comes with extended test coverage beyond default types. Timings will follow.

Plan:
beta != 0 test coverage