@cpuhrsch 2018-12-19
So far we can see up to 3x using the follow script to benchmark

```
import torch
import sys
import time

x = int(sys.argv[1])
y = int(sys.argv[2])
z = int(sys.argv[3])
a = torch.ones(x, y, dtype=torch.int64)
b = torch.ones(y, z, dtype=torch.int64)
c = torch.ones(x, z, dtype=torch.int64)
loops = 0
t1 = time.time()
t2 = 0
while (t2 < 5):
    torch.mm(a, b, out=c)
    t2 = (time.time() - t1)
    loops += 1
t2 = t2 / loops
t2 = int(t2 * 1000 * 1000)
print("loops:\t{}\t x:\t{}\t y:\t{}\t z:\t{}\t time:\t{}ns ".format(str(loops), str(x), str(y), str(z), str(t2)))
```

```
BEFORE:
loops:  2747     x:     128      y:     128      z:     128      time:  1820ns
loops:  120      x:     128      y:     1024     z:     128      time:  41804ns
loops:  325      x:     1024     y:     128      z:     128      time:  15430ns
loops:  16       x:     1024     y:     1024     z:     128      time:  331180ns
loops:  5712     x:     128      y:     128      z:     64       time:  875ns
loops:  246      x:     128      y:     1024     z:     64       time:  20386ns
loops:  685      x:     1024     y:     128      z:     64       time:  7306ns
loops:  31       x:     1024     y:     1024     z:     64       time:  164499ns

AFTER:
loops:  3465     x:     128      y:     128      z:     128      time:  1443ns
loops:  388      x:     128      y:     1024     z:     128      time:  12897ns
loops:  408      x:     1024     y:     128      z:     128      time:  12259ns
loops:  48       x:     1024     y:     1024     z:     128      time:  104480ns
loops:  6634     x:     128      y:     128      z:     64       time:  753ns
loops:  633      x:     128      y:     1024     z:     64       time:  7902ns
loops:  695      x:     1024     y:     128      z:     64       time:  7197ns
loops:  76       x:     1024     y:     1024     z:     64       time:  65889ns
```

This still needs more work:
a) covering the other 3 cases
b) wider coverage of Tensor sizes