@soumith 2018-12-19
does MKL allow you to use int64 sizes via the cblas or mkl API?
If so, ifdef on that would be a good idea as well.