@cpuhrsch 2018-12-21
The rudimentary tilings seems to help quite a bit

```
loops:  4461     x:     128      y:     128      z:     128      time:  1121ns
loops:  505      x:     128      y:     1024     z:     128      time:  9910ns
loops:  506      x:     1024     y:     128      z:     128      time:  9888ns
loops:  61       x:     1024     y:     1024     z:     128      time:  83300ns
loops:  9055     x:     128      y:     128      z:     64       time:  552ns
loops:  824      x:     128      y:     1024     z:     64       time:  6075ns
loops:  902      x:     1024     y:     128      z:     64       time:  5545ns
loops:  97       x:     1024     y:     1024     z:     64       time:  51860ns
```