@cpuhrsch 2019-03-08
Closed in favor of https://github.com/pytorch/pytorch/pull/17730